requirements.txt
llama-index-core
llama-index-llms-google-genai
llama-index-embeddings-google-genai
python-dotenv
